v2/earn_vault_data/python/pyproject.toml
[project]
name = "earn_vault_ranker_python_example"
version = "1.0.0"
description = "Example: Earn Vault Ranker using Compass API"
dependencies = [
    "compass-api-sdk",
    "python-dotenv",
]

